ai_model/summary_model/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Flask==2.3.3
torch==2.6.0+cpu
gunicorn==20.1.0
google-cloud-aiplatform[langchain,reasoningengine]
requests>=2.31.0
tqdm
sentence-transformers
pinecone
python-dotenv>=1.0.0

# cloudpickle==3.0.0
# pydantic==2.9.0
